producer/requirements.txt
confluent-kafka
faker
python-dotenv
